fix: remove braindead enchant score algorithm
This algorithm was already disabled (in a hacky way) by another GearScoreLite
fork, so most people therefore already use GearScore "without enchant support".

I'm disabling it properly and cleanly this time, while leaving a secret,
internal option to enable it manually, in case anyone wants to play with it.

However, the algorithm is batshit insane. The original GearScoreLite author
wasn't particularly skilled, and the old code generated	negative gearscores
rather than simply reducing the	value of unenchanted slots.

It has now been	completely disabled.

Besides, it makes no sense to include enchants in the score, since we cannot
include gems in the score. Because, as the original author noted in a comment
at the top of the file,	"gems are not included in the initial API call".

That might of course just be a consequence of the poor way that GearScoreLite
is written (it doesn't wait for "inspect results ready" events before it tries
to look at the person's items), so the gem issue may be fixable, but honestly
I don't care. By simply looking at the "GearScore without enchants or gems"
we're closer to a proper "ItemLevel Score" addon, which is more in line with
what Blizzard themselves implemented in modern WoW.