fix: remove nonsensical check for combat state
Addons run in a single-threaded context within WoW's main thread, which
means that it makes no sense to check the combat state variable twice,
since NOTHING can change its value between the first time we checked it
at the top of the function, and the second (pointless) time it was being
checked.

Therefore, we'll remove the second, totally pointless check.